fix(api): SI units for every concentration, and a blood group that can be Unknown
Glucose was mg/dL on Vitals, Triage and Stroke and mmol/L on the ambulance
handover, while the band that raises alerts was mg/dL (critical below 50): a
normal 5.4 mmol/L from a meter recorded a critical low. SOFA took bilirubin and
creatinine in mg/dL, the critical-value list mixed mg/dL with mmol/L, nine lab
analytes were US-conventional, and wearable glucose alerts fired at 180/70.

Every concentration is now SI, as South African laboratories report it. The
thresholds are the previous ones converted, not new clinical choices; SOFA uses
the SI bands of the Sepsis-3 table (Singer et al., JAMA 2016, Table 1). The lab
catalogue changes LOINC code with unit, all 30 checked against the NLM service,
which also found that "Free T3" carried the code for total T3. Blood glucose is
stored as a decimal (migration converts existing rows), a critical glucose now
raises a vitals alert (it raised none), and a value that looks like mg/dL is
refused. Toxicology levels move to the scoring catalogue.

BloodType gains Unknown, a Special Case with the same interface as a real
group: registration no longer forces a guess for an untyped patient, and the
loader no longer turns a missing group into O+.

+++ api/src/repositories/traits.rs
@@ -322,7 +322,8 @@ pub struct VitalSignsEntity {
     pub gcs_eye: Option<i32>,
     pub gcs_verbal: Option<i32>,
     pub gcs_motor: Option<i32>,
-    pub blood_glucose: Option<i32>,
+    /// mmol/L.
+    pub blood_glucose: Option<f64>,
     pub weight_kg: Option<f64>,
     pub height_cm: Option<f64>,
     pub bmi: Option<f64>,
@@ -351,7 +352,8 @@ pub struct TriageAssessmentEntity {
     pub oxygen_saturation: Option<i32>,
     pub pain_scale: Option<i32>,
     pub gcs_score: Option<i32>,
-    pub blood_glucose: Option<i32>,
+    /// mmol/L.
+    pub blood_glucose: Option<f64>,
     pub weight: Option<f64>,
     pub is_critical: bool,
     pub requires_isolation: bool,
